ShowsInteractor and delegate
